+++ src/components/map/Map.js
@@ -2,7 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import isEqual from 'lodash/isEqual';
 import pick from 'lodash/pick';
-import { BASEMAPS, LABELS } from 'components/map/constants';
+import { BASEMAPS, LABELS, BOUNDARIES } from 'components/map/constants';
+
+// Redux
+import { connect } from 'react-redux';
 
 // Components
 import Spinner from 'components/ui/Spinner';
@@ -11,6 +14,10 @@ import Spinner from 'components/ui/Spinner';
 /**
  * Basemap
  * @typedef {{ id: string, label: string, value: string, options: any }} Basemap
+ */
+/**
+ * Label
+ * @typedef {{ id: string, label: string, value: string, options: any }} Label
  */
 
 // Leaflet can't be imported on the server because it's not isomorphic
@@ -92,6 +99,10 @@ class Map extends React.Component {
       this.setLabels(nextProps.labels);
     }
 
+    if (this.props.boundaries !== nextProps.boundaries) {
+      this.setBoundaries(nextProps.boundaries);
+    }
+
     // We automatically pan to the bounds if they are provided
     // or updated
     if ((!this.props.mapConfig.bounds && nextProps.mapConfig.bounds)
@@ -142,7 +153,7 @@ class Map extends React.Component {
 
   /**
    * Set the map's basemap
-   * @param {{ id: string, value: string, label: string, options: object }} basemap Basemap
+   * @param {Basemap} basemap Basemap
    */
   setBasemap(basemap) {
     if (this.tileLayer) this.tileLayer.remove();
@@ -153,16 +164,30 @@ class Map extends React.Component {
   }
 
   /**
-   * Toggle the map's labels
-   * @param {boolean} showLabels Whether to show the labels
+   * Set the map's labels
+   * @param {Label} labels Labels
    */
-  setLabels(showLabels) {
-    if (this.labelLayer && !showLabels) this.labelLayer.remove();
+  setLabels(labels) {
+    if (this.labelsLayer) this.labelsLayer.remove();
 
-    if (showLabels) {
-      this.labelLayer = L.tileLayer(LABELS.value, LABELS.options || {})
+    if (labels.id !== 'none') {
+      this.labelsLayer = L.tileLayer(labels.value, labels.options || {})
         .addTo(this.map)
         .setZIndex(this.props.layerGroups.length + 1);
+    }
+  }
+
+  /**
+   * Toggle the visibility of the boundaries on the map
+   * @param {boolean} showBoundaries
+   */
+  setBoundaries(showBoundaries) {
+    if (this.boundariesLayer) this.boundariesLayer.remove();
+
+    if (showBoundaries) {
+      this.boundariesLayer = L.tileLayer(BOUNDARIES.dark.value, BOUNDARIES.dark.options || {})
+        .addTo(this.map)
+        .setZIndex(this.props.layerGroups.length + 2);
     }
   }
 
@@ -262,6 +287,7 @@ class Map extends React.Component {
     // We set the current basemap and labels
     this.setBasemap(this.props.basemap);
     this.setLabels(this.props.labels);
+    this.setBoundaries(this.props.boundaries);
 
     // We add the event listeners
     this.setEventListeners();
@@ -309,12 +335,16 @@ Map.propTypes = {
    * Selected basemap
    * @type {Basemap} basemap
    */
-  basemap: PropTypes.object,
+  basemap: PropTypes.object.isRequired,
   /**
-   * Whether the labels are show or not
-   * @type {boolean} labels
+   * Selected labels
+   * @type {Label} labels
    */
-  labels: PropTypes.bool,
+  labels: PropTypes.object.isRequired,
+  /**
+   * Whether to show the boundaries or not
+   */
+  boundaries: PropTypes.bool.isRequired,
   /**
    * Configuration of the map
    */
@@ -343,9 +373,13 @@ Map.propTypes = {
 };
 
 Map.defaultProps = {
-  basemap: BASEMAPS.dark,
-  labels: false,
   interactionEnabled: true
 };
 
-export default Map;
+const mapStateToProps = ({ widgetEditor }) => ({
+  basemap: BASEMAPS[widgetEditor.basemapLayers.basemap],
+  labels: LABELS[widgetEditor.basemapLayers.labels || 'none'],
+  boundaries: widgetEditor.basemapLayers.boundaries
+});
+
+export default connect(mapStateToProps, null)(Map);