+++ src/helpers/WidgetHelper.js
@@ -693,7 +693,8 @@ export async function getWidgetConfig(
       caption,
       zoom,
       latLng,
-      bounds
+      bounds,
+      basemapLayers
     } = widgetEditor;
 
     let chartConfig = {};
@@ -728,7 +729,8 @@ export async function getWidgetConfig(
         layer_id: layer && layer.id,
         zoom,
         ...latLng,
-        ...bounds && { bbox: [bounds[0][1], bounds[0][0], bounds[1][1], bounds[1][0]] }
+        ...bounds && { bbox: [bounds[0][1], bounds[0][0], bounds[1][1], bounds[1][0]] },
+        basemapLayers
       };
     }